% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{run_Load}
\alias{run_Load}
\title{Read single-cell RNA-seq count matrices in different formats (cellRanger output,
H5 or RDS )
Inputs:
\itemize{
\item input: file/folder path
\item projectName: optional, name for the sample loaded
\item metadata: path to table file containing cell metadata (excel or plain text).
\item cellGroups: optional, variable to separate cells by.
Outputs: SeuratObject containing the experiment's count matrix and cell metadata.
}}
\usage{
run_Load(input, metadata, projectName = "scRNA-seq", cellGroups = NULL)
}
\description{
Read single-cell RNA-seq count matrices in different formats (cellRanger output,
H5 or RDS )
Inputs:
\itemize{
\item input: file/folder path
\item projectName: optional, name for the sample loaded
\item metadata: path to table file containing cell metadata (excel or plain text).
\item cellGroups: optional, variable to separate cells by.
Outputs: SeuratObject containing the experiment's count matrix and cell metadata.
}
}
